+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 ### State Machine Breaking
 
 - [#12](https://github.com/crypto-org-chain/go-ethereum/pull/12) Apply precompile interface change.
+
+### Improvements
+
+- [#](https://github.com/crypto-org-chain/go-ethereum/pull/) Add PrecompiledContractWithChainConfig interface to set the chain config.